+++ src/Tests/Analyzers.Tests/RCS1207ConvertMethodGroupToAnonymousFunctionTests.cs
@@ -596,6 +596,6 @@ class C
         M2(M);
     }
 }
-", options: Options.AddConfigOption(ConfigOptionKeys.UseAnonymousFunctionOrMethodGroup, ConfigOptionValues.UseAnonymousFunctionOrMethodGroup_MethodGroup));
+", options: Options.SetConfigOption(ConfigOptionKeys.UseAnonymousFunctionOrMethodGroup, ConfigOptionValues.UseAnonymousFunctionOrMethodGroup_MethodGroup));
     }
 }